Fix bug where computer gave up if there were no winning moves
If there were no winning moves left the computer correctly did not pick a best move, however this caused the game to stop. It might still be possible for the human player to win, so if there are no winning moves left pick a random empty square and give the move a weight of 0.