Fix bug where game froze if using history after a draw
The game board UI is locked after each human click and then unlocked after the computer go in order to prevent fast clickers taking the computer go. In the event of a draw the computer never got a go so the UI was never unlocked, if the user then rewound the game to a previous state the UI was still locked. Added resetting the UI lock
to the history navigation function.